Add dxgi support
Adds a dxgi dll shim to intercept dxgi API calls.  The core
functionality is moved from the d3d12 dll to a common dx12_capture dll,
which is shared by both dxgi and d3d12 dlls.  The dxgi implementation
includes a custom CreateDXGIFactory2 function and wrappers for the
IDXGIFactory classes.

Change-Id: Idc26c3c672692624a27e31a8ecf2e32e16b264d5