Only wrap objects in calls made by the application
Add a counter to be incremented and decremented when exiting and
entering a function call.  If the count is 1, the function was called by
the application and any objects created by the function should be
wrapped.  If the count is greater than 1, the function was called by
another function in the run time and any objects created by the function
will not be wrapped to prevent wrapped functions from being inserted
into the run time layer.

Change-Id: Ib8990a4734de869f49e115f2b4261c56f0c3895d